feature/tagging: implemented command, requeest, validation, authorization, and write handler (#15)
Note: This is an incremental unit of work. Future PRs on this branch will incorporate tagging into list functionality, lifecycle commands, and find a smart way to store the tag values on the bounty cards.

Implemented command, request, validation, authorization, and write handler

In this PR, tags are upserted values. If/when users find this feature helpful, we can implement an array of tags, as well as a sensible way to remove tags on the bot